fixed tokenAmount being always zero for decimal values

+++ services/payflow-service/src/main/java/ua/sinaver/web3/payflow/message/PaymentUpdateMessage.java
@@ -3,6 +3,6 @@ package ua.sinaver.web3.payflow.message;
 public record PaymentUpdateMessage(Integer chainId, String token, String hash, String fulfillmentId,
                                    Integer fulfillmentChainId,
                                    String fulfillmentHash,
-                                   Integer tokenAmount,
+                                   Double tokenAmount,
                                    String comment) {
 }
